+++ node/src/Commands.ts
@@ -2810,3 +2810,10 @@ export function createLCS(
 
     return createCommand(RequestType.LCS, args);
 }
+
+/**
+ * @internal
+ */
+export function createTouch(keys: string[]): command_request.Command {
+    return createCommand(RequestType.Touch, keys);
+}